Update finduid.sh
Will now pull down all files so you can compare the files you download to the files in the stream. 

Why? What if the sensor has a host with a compromised host / dns and they are being served bs files?

Well....you can just do a tcpdump -i interface and then run this script and download the same files and see if they're coming from the same IP.....and if they are  you can hash them to see if they match the ones in the stream.